backend/requirements.txt
Django==3.2.8
python_bcrypt==0.3.2
requests==2.25.1
